+++ elasticsearch/connection/http_requests.py
@@ -58,4 +58,4 @@ class RequestsHttpConnection(Connection):
 
         self.log_request_success(method, url, response.request.path_url, body, response.status_code, raw_data, duration)
 
-        return response.status_code, raw_data
+        return response.status_code, response.headers, raw_data