+++ Changelog.rst
@@ -13,6 +13,9 @@ Changelog
  * `helpers.bulk` and `helpers.streaming_bulk` are no longer limitted to just
    index operations.
  * unicode body (for `incices.analyze` for example) is now handled correctly
+ * changed `perform_request` on `Connection` classes to return headers as well.
+   This is a backwards incompatible change for people who have developed their own
+   connection class.
 
 0.4.3
 -----